Refactor BatchView page
