+++ src/views/batch/BatchView.test.tsx
@@ -1,4 +1,4 @@
-import { BatchView, addOrdinal, tokenTitle } from "./BatchView";
+import { BatchView, tokenTitle } from "./BatchView";
 import { mockImplicitAccount, mockTezOperation } from "../../mocks/factories";
 import { render, screen, within } from "../../mocks/testUtils";
 import { ghostFA2, ghostTezzard } from "../../mocks/tokens";
@@ -69,20 +69,5 @@ describe("tokenTitle", () => {
 
   it("returns pretty amount if decimals field is present", () => {
     expect(tokenTitle(ghostFA2, "1000321")).toBe("10.00321 Klondike3");
-  });
-});
-
-describe("addOrdinal", () => {
-  it("returns correct value", () => {
-    expect(addOrdinal(1)).toEqual("1st");
-    expect(addOrdinal(2)).toEqual("2nd");
-    expect(addOrdinal(3)).toEqual("3rd");
-    expect(addOrdinal(4)).toEqual("4th");
-    expect(addOrdinal(11)).toEqual("11th");
-    expect(addOrdinal(21)).toEqual("21st");
-    expect(addOrdinal(12)).toEqual("12th");
-    expect(addOrdinal(22)).toEqual("22nd");
-    expect(addOrdinal(13)).toEqual("13th");
-    expect(addOrdinal(23)).toEqual("23rd");
   });
 });
